Add model types in admin actions
