+++ server/actions/admin.ts
@@ -12,6 +12,7 @@ import {
   ICheckTokenInput,
   ITokenPayload
 } from "server/types/admin";
+import { IAdmin } from "utils/types";
 
 const SALT_ROUNDS = 10;
 const JWT_EXPIRES_IN = "1d";
@@ -43,7 +44,7 @@ export async function login({
 }: ILoginInput): Promise<string | never> {
   await Mongo();
 
-  const admin = await Admin.findOne({ email });
+  const admin: IAdmin | null = await Admin.findOne({ email });
   if (admin) {
     if (!(await bcrypt.compare(password, admin.password))) {
       throw new Error(errors.admin.INVALID_LOGIN);
@@ -68,9 +69,9 @@ export async function signup({
     throw new Error(errors.admin.USER_EXISTS);
   }
 
-  const nonprofit = await Nonprofit.exists({ _id: nonprofitId });
+  const nonprofit: boolean = await Nonprofit.exists({ _id: nonprofitId });
   if (nonprofit) {
-    const admin = await Admin.create({
+    const admin: IAdmin = await Admin.create({
       firstName,
       lastName,
       email,
