+++ server/types/models.ts
@@ -1,13 +1,5 @@
 import { Document, Schema } from "mongoose";
 
-export interface IAdmin extends Document {
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  nonprofitId: string;
-}
-
 export interface IDonation extends Document {
   name: string;
   email: string;
